ansible-ncs-clean/pyproject.toml
[project]
name = "ansible-ncs"
version = "0.1.0"
description = "Network & Compute Site Automation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "ansible>=13.2.0",
    "ansible-dev-tools>=26.1.0",
    "ansible-lint>=26.1.1",
    "pip>=26.0.1",
    "pykerberos>=1.2.4",
    "pypandoc>=1.16.2",
    "pyvim>=3.0.3",
    "pyvmomi>=8.0.3.0.1",
    "pywinrm>=0.5.0",
    "requests>=2.32.5",
    "setuptools<82",
    "vsphere-automation-sdk @ git+https://github.com/vmware/vsphere-automation-sdk-python.git@v8.0.3.0",
    "wheel>=0.46.3",
    "yamllint>=1.35.1",
    "ncs-reporter",
]


[tool.uv]
package = false

[tool.uv.sources]
ncs-reporter = { path = "../ncs_reporter", editable = true }

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    "collections/ansible_collections/cloud",
    "collections/ansible_collections/community",
    "collections/ansible_collections/vmware",
    ".venv",
    ".direnv",
]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
check_untyped_defs = true
explicit_package_bases = true
mypy_path = "tools/ncs_reporter/src"
exclude = "collections/ansible_collections/(cloud|community|vmware|windows)/"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
markers = ["integration: ansible-runner integration tests"]

[tool.pyright]
typeCheckingMode = "basic"
pythonVersion = "3.10"
extraPaths = ["tools/ncs_reporter/src"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.venv",
    ".direnv",
    "collections/ansible_collections/cloud",
    "collections/ansible_collections/community",
    "collections/ansible_collections/vmware",
]
reportMissingImports = true
reportMissingTypeStubs = false
